requirements.txt
openai==0.27.9
python_wordpress_xmlrpc==2.3
selenium==4.11.2